Reorder tests from test_server to /routers/test_*
