+++ tests/server/routers/test_links.py
@@ -1,0 +1,11 @@
+# pylint: disable=relative-beyond-top-level
+import unittest
+
+from optimade.models import LinksResponse
+
+from ..test_server import EndpointTests
+
+
+class LinksEndpointTests(EndpointTests, unittest.TestCase):
+    request_str = "/links"
+    response_cls = LinksResponse
